+++ workflow-saga/java/sdk/order-processor-saga/pom.xml
@@ -4,9 +4,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.service</groupId>
-    <artifactId>OrderProcessingSagaService</artifactId>
+    <artifactId>SagaService</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>OrderProcessingSagaService</name>
+    <name>SagaService</name>
     <description>Demo for Dapr workflow Saga</description>
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
@@ -39,7 +39,7 @@
                         </goals>
                         <configuration>
                             <mainClass>
-                                io.dapr.quickstarts.workflows.WorkflowConsoleApp
+                                io.dapr.quickstarts.saga.SagaConsoleApp
                             </mainClass>
                         </configuration>
                     </execution>
