+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/activities/NotifyActivity.java
@@ -1,9 +1,9 @@
-package io.dapr.quickstarts.workflows.activities;
+package io.dapr.quickstarts.saga.activities;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.dapr.quickstarts.workflows.models.Notification;
+import io.dapr.quickstarts.saga.models.Notification;
 import io.dapr.workflows.runtime.WorkflowActivity;
 import io.dapr.workflows.runtime.WorkflowActivityContext;
