+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/activities/ProcessPaymentActivity.java
@@ -1,9 +1,9 @@
-package io.dapr.quickstarts.workflows.activities;
+package io.dapr.quickstarts.saga.activities;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.dapr.quickstarts.workflows.models.PaymentRequest;
+import io.dapr.quickstarts.saga.models.PaymentRequest;
 import io.dapr.workflows.runtime.WorkflowActivity;
 import io.dapr.workflows.runtime.WorkflowActivityContext;
