+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/activities/RequestApprovalActivity.java
@@ -1,10 +1,10 @@
-package io.dapr.quickstarts.workflows.activities;
+package io.dapr.quickstarts.saga.activities;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.dapr.quickstarts.workflows.models.ApprovalResult;
-import io.dapr.quickstarts.workflows.models.OrderPayload;
+import io.dapr.quickstarts.saga.models.ApprovalResult;
+import io.dapr.quickstarts.saga.models.OrderPayload;
 import io.dapr.workflows.runtime.WorkflowActivity;
 import io.dapr.workflows.runtime.WorkflowActivityContext;
