+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/activities/UpdateInventoryActivity.java
@@ -1,4 +1,4 @@
-package io.dapr.quickstarts.workflows.activities;
+package io.dapr.quickstarts.saga.activities;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,10 +6,10 @@ import org.slf4j.LoggerFactory;
 import io.dapr.client.DaprClient;
 import io.dapr.client.DaprClientBuilder;
 import io.dapr.client.domain.State;
-import io.dapr.quickstarts.workflows.models.InventoryItem;
-import io.dapr.quickstarts.workflows.models.InventoryRequest;
-import io.dapr.quickstarts.workflows.models.InventoryResult;
-import io.dapr.quickstarts.workflows.models.OrderPayload;
+import io.dapr.quickstarts.saga.models.InventoryItem;
+import io.dapr.quickstarts.saga.models.InventoryRequest;
+import io.dapr.quickstarts.saga.models.InventoryResult;
+import io.dapr.quickstarts.saga.models.OrderPayload;
 import io.dapr.workflows.runtime.WorkflowActivity;
 import io.dapr.workflows.runtime.WorkflowActivityContext;
