+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/models/ApprovalResult.java
@@ -1,4 +1,4 @@
-package io.dapr.quickstarts.workflows.models;
+package io.dapr.quickstarts.saga.models;
 
 public enum ApprovalResult {
   Unspecified,
