+++ workflow-saga/java/sdk/order-processor-saga/src/main/java/io/dapr/quickstarts/saga/models/OrderPayload.java
@@ -1,4 +1,4 @@
-package io.dapr.quickstarts.workflows.models;
+package io.dapr.quickstarts.saga.models;
 
 public class OrderPayload {
